This PR modifies the type annotation of the `qualifiers` attribute declaration from:

```python
qualifiers: Union[str, Dict[str, str], None]
```

to:

```python
qualifiers: Dict[str, str]
```

in order to reflect its true runtime type.

Closes #169.

### Miscellaneous

- fixed return type of the `encode: "Optional[Literal[False]]"` overload of `normalize_qualifiers` to `Dict[str, str]`
- fixed return type of the `encode: "Optional[Literal[False]]"` overload of `normalize` to `Dict[str, str]` (the other tuple members in the return type unchanged)
- updated instances of `Union[AnyStr, Dict[str, str], None]` to `Optional[Union[AnyStr, Dict[str, str]]]`
- updated instances of `Union[str, Dict[str, str], None]` to `Optional[Union[str, Dict[str, str]]]`
- removed unused `# NOQA` directives
- updated return type of `PackageURL.__new__` and `PackageURL.from_string` to `typing_extensions.Self`
- updated first param name of `PackageURL.__new__` from `self` to `cls` since it is a `classmethod`
- removed quotes from type annotation forward references by adding `from __future__ import annotations`