@tdruez 2024-10-23
@jhoward-lm I see there are still `Union[str, Dict[str, str], None]` references in the source, should those be updated to `Dict[str, str]` as well?