@jhoward-lm 2024-10-23
@tdruez I don't believe so. Those are generalized argument/parameter types, but the type gets narrowed before returning.

You did however draw my attention to several other typing issues; I'll push a commit shortly for review