@jhoward-lm 2024-10-23
@tdruez @gruebel With my most recent comment, tests still pass locally, although the oldest version of python I was able to test with was 3.9 (3.7 and 3.8 are EOL so I couldn't install them with homebrew).

I'll update the PR description with more detail. Let me know your thoughts!