@jhoward-lm 2024-10-23
@gruebel @tdruez 
With the addition of `from __future__ import annotations`, the pipe style union operator as well as built-in collection types (`Dict` -> `dict`, `Tuple` -> `tuple`, etc.) are available. I verified by running `pyupgrade --py37-plus src/packageurl/__init__.py`.

So things like this could be updated across the board:

```python
# before
Union[str, Dict[str, str], None]

# after
str | dict[str, str] | None
```

Let me know if you want me to go ahead with it, or I can hold off.