@jhoward-lm 2024-10-24
@tdruez @gruebel I was able to get python 3.7 installed using `pyenv` to attempt to test, but in case you aren't aware, running `make virtualenv` fails with 

```bash
ImportError: cannot import name 'cached_property' from 'functools'
```