@gruebel 2024-10-24
@jhoward-lm there was no need for testing it with 3.7, because it was dropped with the last release. And CI passed already with your changes, so no need to do extra testing for this PR 😄