@jhoward-lm 2024-10-25
@tdruez @gruebel I threw together a quick [branch](https://github.com/jhoward-lm/packageurl-python/compare/169-qualifiers-typing...lru-cache) on my fork to test caching for faster lookup. Basically repeated calls to `__new__` with the same arguments return the previously constructed `PackageURL` object that was instantiated using those same arguments.

If you're interested, let me know if you want me to create an issue or a draft PR for review.

Tested with this against the current code and after implementing the caching:

```python
if __name__ == "__main__":
    import timeit

    def _test_parsing():
        purl = PackageURL.from_string(
            "pkg:deb/ubuntu/ca-certificates@20230311?arch=all&distro=lunar"
        )

    print("time:", timeit.timeit(stmt=_test_parsing))
```

Current code: ~12.8 seconds
With caching: ~8.5 seconds

The results were a bit more dramatic when I tested with this:

```python
if __name__ == "__main__":
    import timeit

    def _test_parsing():
        purl = PackageURL(
            type="deb",
            namespace="ubuntu",
            name="ca-certificates",
            version="20230311",
            qualifiers="arch=all&distro=lunar",
            subpath=None,
        )

    print("time:", timeit.timeit(stmt=_test_parsing))
```

Current code: ~5.9 seconds
With caching: ~0.5 seconds