@tdruez 2024-10-30
> Are you good with the PR as is?

@jhoward-lm All good, PR merged. Thanks for your contribution 👍